ADD: docker and readme

+++ src/task/task.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { Repository } from 'typeorm';
@@ -28,18 +28,38 @@ export class TaskService {
        skip: (page - 1) * limit,
        take: limit,
      });
+
+
+     if (tasks.length == 0) 
+       throw new NotFoundException(`Tasks not registered`);
+
     return tasks;
   }
 
   async findOne(id: number) {
-    return await this.taskRepository.findOneBy({id});
+    const task = await this.taskRepository.findOneBy({id});
+
+    if (!task) 
+      throw new NotFoundException(`Task with id ${id} not found`);
+
+    return task;
   }
 
   async update(id: number, updateTaskDto: UpdateTaskDto) {
+    const task = await this.taskRepository.findOneBy({id});
+
+    if (!task) 
+      throw new NotFoundException(`Task with id ${id} not found`);
+
     return await this.taskRepository.update(id, updateTaskDto);
   }
 
   async remove(id: number) {
+    const task = await this.taskRepository.findOneBy({ id });
+
+    if (!task) 
+      throw new NotFoundException(`Task with id ${id} not found`);
+
     return await this.taskRepository.delete(id);
   }
 }
